Include a new stub system for eventsub handling
This puts in the basics of the code required for us to be able to listen
for twitch events by including a new subsystem that knows how to handle
that.

Currently this is only stubbed out such that it can provide the
appropriate event handlers to say what it would be doing without
actually doing it, since the code correspondence between this and the
original example is not currently 1:1.