Include setup for all of our EventSub events
This includes the changes required to set up for all of the potential
incoming EventSub events that we're interested in.

To do this we have a large mapping that maps the name of our internal
event to the EventSubListener method that sets up a listener for that
event.

The code will dynamically pull out the methods and invoke them for each
event, setting up a handler that looks for the named event in the
event list and, if found, executes it.