Clarify the event setup and streamline it
Events in Twitch require us to make a request to tell Twitch about a
callback, then wait for a response as Twitch verifies the endpoint.
Once one is set up, until it's removed further setup is much faster
because a quick query can tell what subscriptions need to be added.

Comments have been added which make this process clearer.

Additionally, with as many events as we have out of the box it can take
a while on first auth to set everything up. So, as each call is async
this also modifies it to allow them all to run in parallel to save
time.