Change from one raid event to two
Back when the initial version of the event code that this code is based
on was written, the library that we're using handled From raids and To
raids the same on the back end; as such we ended up just using the From
raid.

In the intervening time the bug was fixed and in fact the two now do
different things.

So, the code here is updated such that we can now distinguish between
an Incoming and an Outgoing raid, and the single raid event has been
adjusted to take that into account.