Include debug logging in all supported events
In order to make it more clear what events mean, what data they provide
and when they trigger, all of the stock events that the system supports
have been updated to provide logging of all of the incoming data in the
event when the event triggers.

To do this, there is a new utility function that can be used to easily
display the properties and methods of the event life as they come in.