Include support for channel point redeems
This includes the initial stubs required for us to support custom
channel point redemptions in the system by having the appropriate db
schema for them, adapters for the handler map, and adding into the
system the ability to find and load all of the custom redeems that
might be in the database.

As of this commit, this is not tied into anything explicitly, so no
redeems will actually fire any code. More work is needed in this area
to get this fully up and running.

At a minimum we need to handle the appropriate events to both capture
new redeems and remove old ones, and do a query for existing redeems
at startup since new items may be added prior to us starting, for which
we might not get an event.