Create a debug method in the event list
This takes a duplicate of the debug code that currently exists in the
utils file, and moves it so that it's inline with the new event list.

The idea is that this version doesn't need to be given information about
the structure of the event because we can infer it from the table, so
all call points will thus be a lot nicer to look at and easier to add
and remove as needed.

To make this more performant, an inside out lookup list is created that
maps our internal names to the list elements that contain them so that
no linear searches are needed at runtime.