Normalize all of the event names
In an act of extreme rebellion, at some point we switched from using
camelCase naming to snake_case naming in events, and once seen it
cannot be unseen.

This normalizes the list to use all snake_case everywhere for event
names.

In theory we could use the twitch name instead, but some events
(most notably the raid event) have multiple possible expressions using
the same name, which makes that problematic.