Use the events from the new list to bootstrap
The bootstrap code has been updated so that instead of having a
redundant list of events that it should add, the list is now
dynamically created from the information in the global event list.

This ensures that when the list of events that we care about is
adjusted, the bootstrap will update accordingly.