Use the events from the new list to get auth scope
Previously the list of scopes to ask for at channel authorization time
was hard-coded in the authorization table.

It seems like a better idea to dynamically grab the list of scopes that
are needed for all known events at startup time by gathering the unique
list of scopes from the event list.

This means that as events are added or removed, any scopes that they
require will be automatically added to the list.