Update the addcmd file to make a better stub
This updates the command that's used to dynamically create a new command
at runtime so that when it generates a new implementation file, it will
populate that file with a more appropriate stub that has the actual
command name.

This means that when we add a new command, it's available right away.