Fix an authentication state oversight
The code that handles the Twitch authorization callback was erroneously
only checking the state parameter to make sure it wasn't being spoofed
in cases where the user actually authenticated.

This means that on any spoof attempt that doesn't contain a valid code,
it would assume you meant to disconnect and would drop the auth even
though the request was spoofed.