Factor the auth provider creation into utils
Previously the helper for creating an auth provider for Twitch was in
the chat code because it was the only thing that actually needed it.

However, one may want to make Twitch requests as the user whose channel
the bot resides in, and in such a case we would actually require a
second instance of the Twitch API that wraps that authorization, so
this code should be made available more globally.