Fix an error with revoking tokens on deauth
The authorization provider that we use is a refreshing provider, which
will automatically try to create a new token when one is needed but the
current one has expired. This refresh only happens on requests that
need a token.

The code that does a deauthorization tries to drop the existing token if
there is one, but the result of this in cases where the token has
expired (for example, if the bot hasn't said anything for an extended
period of time) is that the attempt to revoke the token an error is
raised because it's not actually valid.

This catches such an error and logs it, but otherwise allows the deauth
to continue.