Keep the user authorization token for future use
Previously, the token that was created for the user when they authorize
the bot for the channel was immediately revoked because it was not
needed, as the authorization was only to allow app access to some user
data.

Since we may want to have access to the Twitch API as the user in whose
channel the bot is currently connected, we need to keep the token.

This simple change stops the revocation and makes sure that all tokens
are persisted instead of just the bot token.