Include an event to track when the user API is up
Some code will want to know when the Twitch API attached to the user
token is available so that they can make requests using it.

This sets up a new pair of events which will broadcast when the user API
goes up or comes down, allowing interested code to take appropriate
actions.