Set up an example event that auto-shouts out raids
This includes a simple example of how one could handle incoming events
by reacting to an incoming raid by shouting out the broadcaster who has
raided in.

This is done by "faking" up a command message and then asking the
command to execute it as if the broadcaster manually typed it.

This is commented out but could be easily enabled and extended, such as
to make it delay to ensure that it doesn't conflict with any animated
raid animations for example (which would also be something that you
could manually trigger from here as well).