Include a new subsystem for syncing redeems
This includes the stub of a new subsystem in the bot whose job it is to
synchronize the channel point redeems on the authorized user's Twitch
channel with the list of known redeems that are available locally.

This will allow us to make sure that the list of redeems is up to date
every time the bot starts for cases in which the user has adjusted the
list of redeems while the bot was not running.

More work will need to be done to handle the events that tell us that
channel point redeems are changing so that we can keep the list up to
date at runtime as well.

This will be done by having the core code that takes these actions here
in this file and having the events that are raised call into them as
appropriate.