Include basic stubs for handling redeem changes
This lays the foundational groundwork for being able to synchronize the
list of channel point redeems in the channel with Twitch, both by doing
a request for redeems when authorization is achieved to see what has
changed as well as by handling events that give us information about
changes at runtime.

These are currently only stubbed out and nothing is done as of yet.