Alter how the channel point monitor dispatch works
The original idea was that there would be a single entry point which
could be given a single channel point redeem detail and determine what
to do with it by invoking the appropriate handler.

In retrospect there's no good way to know just from a single item
whether or not it represents an addition or a deletion without knowing
the whole scope of the existing list.

So, here we're having the channel point events directly invoke the
handler for the type of operation they know it must be, since that will
always be operating from a known state.

The code for the startup of the monitor will be adjusted to be more
smarterer.