Include a folder for holding channel point redeems
Unlike the command and event folders which will always exist in the code
base because there are defaults that always exist, for channel point
redeems that doesn't make a lot of sense.

So, we need to ensure that this folder actually exists so that at
runtime it doesn't get all angry at us.