Include a stub channel point redemption file
Include in the stubs folder a stub that can be used to generate a new
channel point redemption handler. This is done in the same style as the
one we recently added for command stubs.

The stub displays the event information for the redeem event that
invoked it as a debug item, so the channel point redeem handler has
it's debug logic commented out.

The idea is that the channel point redeem handler is fully functional as
intended for the example, so the extra debug information is not needed
there.