Fix the editor for situations with no files
Although it can't happen for commands and events because there are
packed in examples of both, it's possible for there to be no channel
point redeem entries in the database.

In such a case the editor dropdown will have no items, which will leave
the editor in an inconsistent state because it shows the last loaded
content and gives the impression that you can reload it when you
cannot.

This simple change makes sure that if, on selection, there are no items
of the appropriate type, the discard button is disabled and the file
buffer is cleared and set to be read only.

The button will be re-enabled on the next load operation (if any).