Flesh out the handling for doing initial refresh
This updates the simple stub code which was only exercising that we know
how to talk to the Twitch API so that it will gather the list of known
redeems both from Twitch and from the Database and determine which
items are missing from either list.

From this it can infer which items need to be added to the DB because
they exist on Twitch and which items should be removed because they are
no longer present in the channel.

For any items that exist in both lists, run them through the handler
that allows us to see if their names have changed so that they can be
updated locally as appropriate.

Although it may seem wasteful to not double check at this point that the
titles are the same, the update events we get at runtime will be
generated for any change, not just titles, so the code needs to be
smart enough to handle this situation anyway.