Implement the ability to add dynamic redeems
This implements the basic initial version of the handler which, given
the unique Twitch ID and human readable name of a channel point redeem
in a particular channel will update the database to include such an
item and create a stub file on disk that implements it.

In the process some of the API's we just implemented needed to change as
we require the broadcaster's userID in order to ensure that our titles
are unique.