Implement the ability to remove dynamic redeems
This implements the initial version of code that can, when told that a
redeem no longer exists on Twitch, remove it from the database and
unload its file, making it no longer active.

When removed, the implementation file remains behind so that it can be
put back in place if the redeem returns.