Change the semantics of channel redeem updates
This adds a couple of small but important changes to the code that does
channel point redemption updating.

First, when gathering the list of items that exist in both the DB and on
Twitch in order to see if they're named, we need to be grabbing the
exemplar items out of the Twitch Dataset instead of the DB dataset;
otherwise we would end up trying to compare the database name to itself
and deciding that nothing can ever be changed.

Secondly, it's important to always delete defunct items before we add
new ones because it seems cleaner to sweep away older redeems (say from
a different channel if the auth changed) before adding in the new ones.