Update the helper for making an entry title
In order to make sure that channel point redeem titles are unique in the
database, we have a helper that attaches the user's broadcaster ID to
to the end of the title.

Previously this would always put that suffix on the end. However in a
situation where the title in question already has it, this would double
it up, which is undesirable.

Although this should not happen any longer in practice (noticing this
happening is what prompted the prior change to make sure we always do
renames from the Twitch data set) it seems like a good idea to put this
in place anyway.