Implement the final changes for channel redeems
This finalizes the last portion of the channel point redemption logic,
which allows us to detect when existing channel point entries have been
renamed and update their names in the database to keep things in sync.

This is entirely a cosmetic change, but it is a useful one because it
makes sure that the names that are seen in the user interface match the
current reality.