Change the default command prefix to ! from $
During testing and implementation of the bot in the channel, the default
command prefix on everything was $ so that it doesn't collide with other
commands in the existing bots.

Since we're approaching the end of the project timeline, change that to
an appropriate default command prefix for more public consumption
instead.